+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -6184,29 +6184,31 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                                 context=self.REQUEST)
                 self.plone_utils.addPortalMessage(msg)
 
-    def _createOrUpdatePloneGroup(self, groupSuffix, groupTitleSuffix=None, only_group_ids=False):
+    def _createOrUpdatePloneGroup(self, groupSuffix, groupTitleSuffix=None, dry_run_return_group_ids=False):
         '''Create a group for this MeetingConfig using given p_groupSuffix to manage group id and group title.
            This will return groupId and True if group was added, False otherwise.'''
         groupId = "{0}_{1}".format(self.getId(), groupSuffix)
-        if only_group_ids:
+        if dry_run_return_group_ids:
             return groupId, False
         groupTitle = self.Title(include_config_group=True)
         if groupTitleSuffix:
             groupSuffix = safe_unicode(groupTitleSuffix)
-        wasCreated = createOrUpdatePloneGroup(groupId=groupId, groupTitle=groupTitle, groupSuffix=groupSuffix)
+        wasCreated = createOrUpdatePloneGroup(
+            groupId=groupId, groupTitle=groupTitle, groupSuffix=groupSuffix)
         return groupId, wasCreated
 
     security.declarePrivate('createPowerObserversGroups')
 
-    def createPowerObserversGroups(self, force_update_access=False, only_group_ids=False):
+    def createPowerObserversGroups(self, force_update_access=False, dry_run_return_group_ids=False):
         '''Creates Plone groups to manage power observers.'''
         groupIds = []
         tool = api.portal.get_tool('portal_plonemeeting')
         for po_infos in self.getPowerObservers():
             groupSuffix = po_infos['row_id']
-            groupId, wasCreated = self._createOrUpdatePloneGroup(groupSuffix,
-                                                                 groupTitleSuffix=po_infos['label'],
-                                                                 only_group_ids=only_group_ids)
+            groupId, wasCreated = self._createOrUpdatePloneGroup(
+                groupSuffix,
+                groupTitleSuffix=po_infos['label'],
+                dry_run_return_group_ids=dry_run_return_group_ids)
             groupIds.append(groupId)
             if wasCreated or force_update_access:
                 # now define local_roles on the tool so it is accessible by this group
@@ -6224,25 +6226,27 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
     security.declarePrivate('createBudgetImpactEditorsGroup')
 
-    def createBudgetImpactEditorsGroup(self, only_group_ids=False):
+    def createBudgetImpactEditorsGroup(self, dry_run_return_group_ids=False):
         '''Creates a Plone group that will be used to apply the 'MeetingBudgetImpactEditor'
            local role on every items of this MeetingConfig regarding self.itemBudgetInfosStates.'''
         groupIds = []
-        groupId, wasCreated = self._createOrUpdatePloneGroup(groupSuffix=BUDGETIMPACTEDITORS_GROUP_SUFFIX,
-                                                             only_group_ids=only_group_ids)
+        groupId, wasCreated = self._createOrUpdatePloneGroup(
+            groupSuffix=BUDGETIMPACTEDITORS_GROUP_SUFFIX,
+            dry_run_return_group_ids=dry_run_return_group_ids)
         groupIds.append(groupId)
         return groupIds
 
     security.declarePrivate('createMeetingManagersGroup')
 
-    def createMeetingManagersGroup(self, force_update_access=False, only_group_ids=False):
+    def createMeetingManagersGroup(self, force_update_access=False, dry_run_return_group_ids=False):
         '''Creates a Plone group that will be used to apply the 'MeetingManager'
            local role on every plonemeeting folders of this MeetingConfig and on this MeetingConfig.'''
         groupIds = []
-        groupId, wasCreated = self._createOrUpdatePloneGroup(groupSuffix=MEETINGMANAGERS_GROUP_SUFFIX,
-                                                             only_group_ids=only_group_ids)
+        groupId, wasCreated = self._createOrUpdatePloneGroup(
+            groupSuffix=MEETINGMANAGERS_GROUP_SUFFIX,
+            dry_run_return_group_ids=dry_run_return_group_ids)
         groupIds.append(groupId)
-        if not only_group_ids and wasCreated or force_update_access:
+        if not dry_run_return_group_ids and wasCreated or force_update_access:
             # now define local_roles on the tool so it is accessible by this group
             tool = api.portal.get_tool('portal_plonemeeting')
             tool.manage_addLocalRoles(groupId, ('MeetingManager',))
@@ -6258,13 +6262,14 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
     security.declarePrivate('createItemTemplateManagersGroup')
 
-    def createItemTemplateManagersGroup(self, force_update_access=False, only_group_ids=False):
+    def createItemTemplateManagersGroup(self, force_update_access=False, dry_run_return_group_ids=False):
         '''Creates a Plone group that will be used to store users able to manage item templates.'''
         groupIds = []
-        groupId, wasCreated = self._createOrUpdatePloneGroup(groupSuffix=ITEMTEMPLATESMANAGERS_GROUP_SUFFIX,
-                                                             only_group_ids=only_group_ids)
+        groupId, wasCreated = self._createOrUpdatePloneGroup(
+            groupSuffix=ITEMTEMPLATESMANAGERS_GROUP_SUFFIX,
+            dry_run_return_group_ids=dry_run_return_group_ids)
         groupIds.append(groupId)
-        if not only_group_ids and wasCreated or force_update_access:
+        if not dry_run_return_group_ids and wasCreated or force_update_access:
             # now define local_roles on the tool so it is accessible by this group
             tool = api.portal.get_tool('portal_plonemeeting')
             tool.manage_addLocalRoles(groupId, (READER_USECASES[ITEMTEMPLATESMANAGERS_GROUP_SUFFIX],))
@@ -6277,26 +6282,34 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             self.itemtemplates.manage_addLocalRoles(groupId, ('Manager', ))
         return groupIds
 
-    def _createOrUpdateAllPloneGroups(self, force_update_access=False, only_group_ids=False):
+    def _createOrUpdateAllPloneGroups(self, force_update_access=False, dry_run_return_group_ids=False):
         """Create or update every linked Plone groups.
            If p_force_update_access this will force update of access given to created group.
-           If p_only_group_ids, this will not create groups but return group ids that would be created."""
+           If p_dry_run_return_group_ids=True, this will not create groups but return
+           group ids that would be created."""
         group_ids = []
         # Create the corresponding group that will contain MeetingManagers
         group_ids += self.createMeetingManagersGroup(
             force_update_access=force_update_access,
-            only_group_ids=only_group_ids)
+            dry_run_return_group_ids=dry_run_return_group_ids)
         # Create the corresponding group that will contain item templates Managers
         group_ids += self.createItemTemplateManagersGroup(
             force_update_access=force_update_access,
-            only_group_ids=only_group_ids)
+            dry_run_return_group_ids=dry_run_return_group_ids)
         # Create the corresponding group that will contain MeetingBudgetImpactEditors
-        group_ids += self.createBudgetImpactEditorsGroup(only_group_ids=only_group_ids)
+        group_ids += self.createBudgetImpactEditorsGroup(
+            dry_run_return_group_ids=dry_run_return_group_ids)
         # Create the corresponding group that will contain MeetingPowerObservers
         group_ids += self.createPowerObserversGroups(
             force_update_access=force_update_access,
-            only_group_ids=only_group_ids)
+            dry_run_return_group_ids=dry_run_return_group_ids)
+        group_ids += self.adapted()._custom_createOrUpdateGroups(
+            force_update_access=force_update_access, dry_run_return_group_ids=dry_run_return_group_ids)
         return group_ids
+
+    def _custom_createOrUpdateGroups(self, force_update_access=False, dry_run_return_group_ids=False):
+        '''See doc in interfaces.py.'''
+        return []
 
     def _set_default_faceted_search(self, collection_id='searchmyitems'):
         """ """
@@ -7564,7 +7577,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
     def displayGroupsAndUsers(self):
         """Display groups and users specific to this MeetingConfig (meetingmanagers, powerobservers, ...)."""
-        plone_group_ids = self._createOrUpdateAllPloneGroups(only_group_ids=True)
+        plone_group_ids = self._createOrUpdateAllPloneGroups(dry_run_return_group_ids=True)
         # include also group "Administrators"
         plone_group_ids.append("Administrators")
         portal = api.portal.get()